requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch
opencv-python
numpy
pandas
tqdm
scikit-learn
open-clip-torch
matplotlib
seaborn
Pillow
